feat: also use reliable storage for plugin configs
