+++ Dalamud/Configuration/Internal/DalamudConfiguration.cs
@@ -436,6 +436,10 @@ internal sealed class DalamudConfiguration : IServiceType, IDisposable
             {
                 deserialized =
                     JsonConvert.DeserializeObject<DalamudConfiguration>(text, SerializerSettings);
+                
+                // If this reads as null, the file was empty, that's no good
+                if (deserialized == null)
+                    throw new Exception("Read config was null.");
             });
         }
         catch (FileNotFoundException)
